backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
googlemaps==4.10.0
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0



